Do not install memcached if it's already installed

+++ cookbooks/memcached/recipes/install_from_source.rb
@@ -2,7 +2,7 @@ memcached_version = node['memcached']['version']
 memcached_download_url = node['memcached']['download_url']
 memcached_installer_directory = '/opt/memcached-installer'
 
-Chef::Log.info "Installing memcached #{memcached_version} from source..."
+Chef::Log.info "Installing memcached #{memcached_version} from source"
 
 # required when installing memcached from source
 package "libevent-dev"
@@ -28,22 +28,28 @@ remote_file "/opt/memcached-#{memcached_version}.tar.gz" do
   backup 0
 end
 
-execute "unarchive Memcached installer" do
-  cwd "/opt"
-  command "tar zxf memcached-#{memcached_version}.tar.gz && sync"
-end
+memcached_installed_version = Mixlib::ShellOut.new 'memcached --version'
+memcached_installed_version.run_command
+if memcached_installed_version.stdout.chomp == "memcached #{memcached_version}"
+  Chef::Log.info "memcached #{memcached_version} is already installed. Skipping installation"
+else
+  execute "unarchive Memcached installer" do
+    cwd "/opt"
+    command "tar zxf memcached-#{memcached_version}.tar.gz && sync"
+  end
 
-execute "Remove old memcached-installer" do
-  command "rm -rf /opt/memcached-installer"
-end
+  execute "Remove old memcached-installer" do
+    command "rm -rf /opt/memcached-installer"
+  end
 
-execute "rename /opt/memcached-#{memcached_version} to /opt/memcached-installer" do
-  command "mv /opt/memcached-#{memcached_version} #{memcached_installer_directory}"
-end
+  execute "rename /opt/memcached-#{memcached_version} to /opt/memcached-installer" do
+    command "mv /opt/memcached-#{memcached_version} #{memcached_installer_directory}"
+  end
 
-execute "run memcached-installer/configure, make, install" do
-  cwd memcached_installer_directory
-  command "./configure && make && make install"
+  execute "run memcached-installer/configure, make, install" do
+    cwd memcached_installer_directory
+    command "./configure && make && make install"
+  end
 end
 
 cookbook_file "/etc/systemd/system/memcached.service" do
